re-organised certin files, the fractal shader into its own component and twearking the visualizer css to adjust properly on window size change.
