+++ src/renderer/src/assets/index.css
@@ -2,15 +2,17 @@
 @tailwind components;
 @tailwind utilities;
 
-canvas {
-  height: 100%;
-  width: 100%;
-  padding: 0;
-  margin: 0;
+* {
+  min-width: 0;
 }
+html {
+  height: 100%;
+}
+#root {
+  height: 100%;
+}
+
 body {
-  display: flex;
-  flex-direction: column;
   font-family:
     Roboto,
     -apple-system,
@@ -22,7 +24,7 @@ body {
     'Cantarell',
     'Open Sans',
     sans-serif;
-  color: #86a5b1;
+  height: 100%;
   background-color: #2f3241;
 }
