+++ src/renderer/src/components/Visualizer.jsx
@@ -3,11 +3,7 @@ import React, { useMemo, useState } from 'react';
 import Scene from './Scene';
 
 function Visualizer() {
-  return (
-    <div className="bg-black mx-4 h-auto">
-      <Scene />
-    </div>
-  );
+  return <Scene />;
 }
 
 export default Visualizer;
