+++ src/renderer/src/components/menus/AudioControls.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function AudioControls() {
+  return (
+    <div>
+      <input></input>
+    </div>
+  );
+}
+
+export default AudioControls;
